feat: refactor code + folder structure
