+++ package.json
@@ -27,7 +27,7 @@
         "dev": "react-scripts start",
         "build": "react-scripts build",
         "lint": "npx eslint .",
-        "lint:fix": "npx eslint --fix && yarn format",
+        "lint:fix": "npx eslint --fix . && yarn format",
         "format": "prettier --write .",
         "test": "react-scripts test",
         "eject": "react-scripts eject",
