+++ src/components/Home/Header/index.js
@@ -2,15 +2,16 @@ import React from 'react';
 import { NavLink } from 'react-router-dom';
 import { Link } from 'react-scroll';
 
-import dark_comp_name from '../../../../assets/images/dark_comp_name.svg';
-import dark_mode_img from '../../../../assets/images/dark_mode_img.svg';
-import darkModeImg from '../../../../assets/images/darkModeImg.svg';
-import darkmodelogo from '../../../../assets/images/darkmodelogo.png';
-import instaLogo from '../../../../assets/images/InstaLogo.svg';
-import CompName from '../../../../assets/images/Instaraise.svg';
-import { HEADER_DATA } from '../../../../config/HomeConfig/HeaderConfig/config.header';
-import { HEADER_SOCIAL_IMAGE } from '../../../../config/HomeConfig/HeaderConfig/config.header';
-import { HEADER_SOCIAL_IMAGE2 } from '../../../../config/HomeConfig/HeaderConfig/config.header';
+import dark_comp_name from '../../../assets/images/dark_comp_name.svg';
+import dark_mode_img from '../../../assets/images/dark_mode_img.svg';
+import darkModeImg from '../../../assets/images/darkModeImg.svg';
+import darkmodelogo from '../../../assets/images/darkmodelogo.png';
+import instaLogo from '../../../assets/images/InstaLogo.svg';
+import CompName from '../../../assets/images/Instaraise.svg';
+import { HEADER_SOCIAL_IMAGE } from '../../../config/HomeConfig/HeaderConfig/config.header';
+import { HEADER_SOCIAL_IMAGE2 } from '../../../config/HomeConfig/HeaderConfig/config.header';
+import { HEADER_DATA } from '../../../config/HomeConfig/HeaderConfig/config.header';
+
 const Header = ({ except }) => {
     var show_dark_header_images = (
         <>
