--- src/config/HomeConfig/StatsConfig/config.stats.js
@@ -1,44 +1,0 @@
-import BurnedImg from '../../../assets/images/burned.svg';
-import Icon from '../../../assets/images/Icon.svg';
-import MarketCapImg from '../../../assets/images/market.svg';
-import PriceImg from '../../../assets/images/price.svg';
-import CirculatingSupplyImg from '../../../assets/images/supply.svg';
-import tvl_img from '../../../assets/images/TVL.svg';
-export const PRICE_DATA = [
-    {
-        id: 1,
-        name: 'Token ticker',
-        description: '$INSTA',
-        image: PriceImg,
-    },
-    {
-        id: 2,
-        name: 'Price',
-        description: '$0.007',
-        image: Icon,
-    },
-    {
-        id: 3,
-        name: 'Market cap',
-        description: '$109,613',
-        image: MarketCapImg,
-    },
-    {
-        id: 4,
-        name: 'Circulating supply',
-        description: '14,652,871',
-        image: CirculatingSupplyImg,
-    },
-    {
-        id: 5,
-        name: 'Burned',
-        description: '$75572.97',
-        image: BurnedImg,
-    },
-    {
-        id: 6,
-        name: 'TVL',
-        description: '$30,087',
-        image: tvl_img,
-    },
-];
