--- src/config/HomeConfig/TrendingConfig/config.trending.js
@@ -1,53 +1,0 @@
-import rightarrow from '../../../assets/images/rightarrow.svg';
-export const TRENDNEWS_DATA = [
-    {
-        id: 1,
-        image_url:
-            'https://blog.instaraise.io/wp-content/uploads/2022/06/3.jpg',
-        subtitle: 'Understanding Decentralized Finance (DeFi) Evolution',
-        category: ['DefiTools', 'All'],
-        descHeader:
-            'DeFi, short for “Decentralized Finance” has been among the hottest topics in the past couple of years. In this post, we will discover the evolution of Decentralized finance and also explore what is hidden in the future for us.',
-        description:
-            'DeFi is one of the revolutionary applications of cryptocurrency and blockchain technology. It’s an alternative to traditional finance without the involvement of third-party.',
-        descPara:
-            'Due to its potential to disrupt the existing status quo in global financial ecosystems, DeFi and DeFi-based applications are slated to provide limitless opportunities to the global population while forcing their traditional counterparts to innovate.',
-        descPara2:
-            'To realize the importance of DeFi, it is necessary to understand the underlying principles of DeFi and how it is different from the more traditional, centralized financial ecosystem that has been in use for a long time.',
-        cardDate: '27 June, 2022',
-        arrowImg: rightarrow,
-    },
-    {
-        id: 2,
-        image_url:
-            'https://blog.instaraise.io/wp-content/uploads/2022/06/multichain_2.jpg',
-        subtitle:
-            'Potential of Multichain in DeFi Ecosystem- Instaraise Spaces Recap',
-        category: ['DefiTools', 'All'],
-        description:
-            'What’s ahead for the the future of Web3 and DeFi ecosystem? Founders of Instaraise and ShuttleOne discusses the advantages of Multichain and explore the possibilities of including Metachain support in InstaDEX',
-        descPara:
-            'Missed our episode of Instaraise Spaces on Twitter? Don’t fret, we have made notes of everything significant that happened during the session. If you have time, you are welcome to listen to the recording of the entire conversation here.',
-        descPara2:
-            'The latest edition of Instaraise Spaces, a series of interactive sessions between the Instaraise team and the larger crypto community was centered around the importance of multichain capabilities in today’s DeFi ecosystem. Along with Instaraise co-founders Dibyo Majumder and Adita Gautam, Hong Zhuang L, founder, and CEO of ShuttleOne discussed the need for interoperability between different protocols and how the respective projects are addressing the crypto community’s needs.',
-        cardDate: '04 June, 2022',
-        arrowImg: rightarrow,
-    },
-    {
-        id: 3,
-        image_url:
-            'https://blog.instaraise.io/wp-content/uploads/2021/12/token-sale-recap.jpg',
-        subtitle:
-            'ShuttleOne $SZO Token sale recap: Bringing the award-winning company',
-        category: ['DevUpdates', 'All'],
-        description:
-            'The first-ever decentralized fundraising and incubation protocol on Tezos Ecosystem, Instaraise has successfully concluded another token offering on its IDO platform. This time, it is the ShuttleOne Network — an operating system for digital finance that enables multichain functionalities to dApps on Metachain.',
-        descPara:
-            'Organized as a one-day event on December 12, 2021, the 1,000,000 $SZO token offering on Instaraise received a great response from the community as expected. In no time, the participants previously whitelisted for the token sale lapped up over $64,000 worth of ShuttleOne’s Tezos-based $SZO utility tokens to turn the sale into a grand success.',
-        descPara2:
-            'The successful ShuttleOne IDO has added another feather to Instaraise’s cap, which already has multiple successes under its belt.',
-
-        cardDate: '21 Dec, 2022',
-        arrowImg: rightarrow,
-    },
-];
