+++ src/container/home/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
-import HomeComponent from '../../components/home/HomeComponent';
+import HomeComponent from '../../components/Home';
+
 const HomeContainer = () => {
     return <HomeComponent />;
 };
